\name{RSKC-internal}
\title{Internal RSKC functions}
\alias{print.rskc}
\alias{print.summary.rskc}
\alias{summary.rskc}
\description{Internal RSKC functions}
\usage{
\method{print}{rskc}(x,...)
\method{summary}{rskc}(object,...)
\method{print.summary}{rskc}(x,...)
}
\author{Yumi Kondo <y.kondo@stat.ubc.ca>}
\keyword{internal}




