\name{RSKC-internal}
\title{Internal RSKC functions}
\alias{print.rskc}
\alias{print.summary.rskc}
\alias{summary.rskc}
\description{Internal RSKC functions}
\usage{
print.rskc(x,...)
summary.rskc(object,...)
print.summary.rskc(x,...)
}
\author{Yumi Kondo <y.kondo@stat.ubc.ca>}
\keyword{internal}




